+++ src/test/java/net/sf/jabref/logic/cleanup/MoveFilesCleanupTest.java
@@ -1,0 +1,60 @@
+package net.sf.jabref.logic.cleanup;
+
+import net.sf.jabref.*;
+import net.sf.jabref.logic.journals.JournalAbbreviationLoader;
+import net.sf.jabref.logic.journals.JournalAbbreviationRepository;
+import net.sf.jabref.model.database.BibDatabase;
+import net.sf.jabref.model.entry.BibEntry;
+import net.sf.jabref.model.entry.FileField;
+import net.sf.jabref.model.entry.ParsedFileField;
+import org.junit.Assert;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.Test;
+import org.junit.rules.TemporaryFolder;
+
+import java.io.File;
+import java.io.IOException;
+
+import static org.junit.Assert.*;
+import static org.mockito.Mockito.mock;
+
+public class MoveFilesCleanupTest {
+
+    @Rule
+    public TemporaryFolder bibFolder = new TemporaryFolder();
+
+    private File pdfFolder;
+    private BibDatabaseContext databaseContext;
+    private MoveFilesCleanup cleanup;
+
+    @Before
+    public void setUp() throws IOException {
+        Globals.prefs = JabRefPreferences.getInstance();
+
+        pdfFolder = bibFolder.newFolder();
+        MetaData metaData = new MetaData();
+        metaData.setDefaultFileDirectory(pdfFolder.getAbsolutePath());
+        metaData.setFile(bibFolder.newFile("test.bib"));
+        databaseContext = new BibDatabaseContext(new BibDatabase(), metaData, new Defaults());
+
+        cleanup = new MoveFilesCleanup(databaseContext);
+    }
+
+    @Test
+    public void movesFileFromSubfolder() throws IOException {
+        File subfolder = bibFolder.newFolder();
+        File tempFile = new File(subfolder, "test.pdf");
+        tempFile.createNewFile();
+        assertTrue(new File(subfolder, "test.pdf").exists());
+
+        BibEntry entry = new BibEntry();
+        ParsedFileField fileField = new ParsedFileField("", tempFile.getAbsolutePath(), "");
+        entry.setField("file", FileField.getStringRepresentation(fileField));
+
+        cleanup.cleanup(entry);
+
+        assertFalse(new File(subfolder, "test.pdf").exists());
+        assertTrue(new File(pdfFolder, "test.pdf").exists());
+    }
+}